requirements.txt
asn1crypto==1.4.0
attrs==22.1.0
bcrypt==3.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==3.0.4
idna==2.10
iniconfig==1.1.1
mnemonic==0.20
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pycryptodome==3.9.9
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.3
requests==2.24.0
six==1.16.0
tomli==2.0.1
urllib3==1.25.11
